+++ src/index.tsx
@@ -4,6 +4,8 @@ export * from './config';
 export * from './types';
 export * from './utils';
 
+export * from './hooks';
+
 export * from './atomics';
 export * from './molecules';
 export * from './organisms';