+++ src/organisms/Filter/components/Body/index.tsx
@@ -1,0 +1,9 @@
+import { FilterBodyWrapper } from './styles';
+
+export const FilterBody = () => {
+    return (
+        <FilterBodyWrapper>
+            {/* <Input.Filter placeholder="Filter for user by email address or username" /> */}
+        </FilterBodyWrapper>
+    );
+};