+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0]
+
+### Add
+
+-   Add `Filter` component.
+-   Add `Tag.Group` component.
+-   Add `type` prop for `Tag` component.
+-   Add  `useFilter` hooks.
+
 ## [1.10.1]
 
 ### Fixed